Database/requirements.txt
yfinance==0.2.27
pandas
numpy
python-dotenv
